Hibernate/Database update:
Therapists (and by extension UserAccounts), Objectives and Steps can now be added to the database. This currently only happens with single example cases of testTherapist, testObjective and testStep objects at program initialization within Controller.java.

Changes to Child and Guardian are only minor, inconsequential changes, to get ready for the next stage of Hibernate implementation, which will allow every major object to be added to the database.